split rust codegen into multiple modules
